[lib] Split types
