+++ packages/lib/src/classes/ChatCompletionService.ts
@@ -4,10 +4,13 @@ import { ConversationRequestOptions } from "./ConversationRequestOptions.js";
 import { ConversationHistory } from "./ConversationHistory.js";
 import { ConversationCallableFunctions } from "./ConversationCallableFunctions.js";
 import { ConversationRequestOptionsModel } from "../schemas/conversationRequestOptions.schema.js";
-import { HandleChatCompletionOptions, PromptOptions } from "../utils/types.js";
 import { ModerationException } from "../exceptions/ModerationException.js";
 import createDryChatCompletion from "../utils/createDryChatCompletion.js";
 import createChatCompletion from "../utils/createChatCompletion.js";
+import {
+    HandleChatCompletionOptions,
+    PromptOptions,
+} from "../utils/types/index.js";
 
 /**
  * Encapsulates the logic for sending requests to the OpenAI API for the Create Chat Completion endpoint.
