+++ packages/lib/src/utils/index.ts
@@ -2,4 +2,4 @@ export { default as createChatCompletion } from "./createChatCompletion.js";
 export { default as createDryChatCompletion } from "./createDryChatCompletion.js";
 export { default as getRequestHeaders } from "./getRequestHeaders.js";
 export { default as getRequestUrl } from "./getRequestUrl.js";
-export * from "./types.js";
+export * from "./types/index.js";
