+++ packages/lib/src/utils/types/callableFunction.types.ts
@@ -1,0 +1,10 @@
+import { JsonSchemaObject } from "../../schemas/jsonSchema.schema.js";
+
+/**
+ * A function that can be called by the model.
+ */
+export interface CreateChatCompletionFunction {
+    name: string;
+    description?: string;
+    parameters?: JsonSchemaObject;
+}
