+++ packages/lib/src/utils/types/chatCompletionService.types.ts
@@ -1,17 +1,9 @@
-/* eslint-disable @typescript-eslint/no-unused-vars */
-import { Conversation } from "../classes/Conversation.js";
-import { ConversationConfig } from "../classes/ConversationConfig.js";
-import { Message } from "../classes/Message.js";
-import { JsonSchemaObject } from "../schemas/jsonSchema.schema.js";
-
-/**
- * Supported values for the `role` property of a message.
- */
-export type ChatCompletionRequestMessageRoleEnum =
-    | "user"
-    | "system"
-    | "assistant"
-    | "function";
+import { CreateChatCompletionFunction } from "./callableFunction.types.js";
+import {
+    CreateChatCompletionFunctionCallMessage,
+    CreateChatCompletionFunctionMessage,
+    CreateChatCompletionMessage,
+} from "./message.types.js";
 
 /**
  * Overridable {@link CreateChatCompletionRequest} properties of a {@link Conversation}'s config for a single prompt.
@@ -25,108 +17,6 @@ export type PromptOptions = Omit<
  * Overridable {@link CreateChatCompletionRequest} properties of a {@link Conversation}'s config for a chat completion.
  */
 export type HandleChatCompletionOptions = Omit<PromptOptions, "stream">;
-
-/**
- * Listener for when a {@link Message} is added to a conversation.
- */
-export type AddMessageListener = (
-    /**
-     * The {@link Message message} that was added
-     */
-    message: Message
-) => void;
-
-/**
- * Listener for when a {@link Message} is removed from a conversation.
- */
-export type RemoveMessageListener = (
-    /**
-     * The {@link Message message} that was removed
-     */
-    message: Message
-) => void;
-
-/**
- * A message in OpenAI's chat format.
- *
- * @see {@link https://platform.openai.com/docs/api-reference/chat/create#chat/create-messages Create Chat Completion Request Body - messages}
- */
-export interface CreateChatCompletionMessage {
-    /**
-     * The message role.
-     *
-     * @see {@link ChatCompletionRequestMessageRoleEnum}
-     */
-    role: Exclude<ChatCompletionRequestMessageRoleEnum, "function">;
-
-    /**
-     * The message content.
-     */
-    content: string;
-
-    function_call?: undefined;
-    name?: undefined;
-}
-
-export type CompletionMessage = Message & {
-    role: Exclude<ChatCompletionRequestMessageRoleEnum, "function">;
-    content: string;
-    functionCall: undefined;
-    name: undefined;
-};
-
-/**
- * A function_call-related message in OpenAI's chat format.
- *
- * @see {@link https://platform.openai.com/docs/api-reference/chat/create#chat/create-messages Create Chat Completion Request Body - messages}
- */
-export interface CreateChatCompletionFunctionCallMessage {
-    role: Extract<ChatCompletionRequestMessageRoleEnum, "assistant">;
-    content: null;
-    function_call: {
-        name: string;
-        arguments: string;
-    };
-
-    name?: undefined;
-}
-
-export type FunctionCallMessage = Message & {
-    role: "assistant";
-    content: null;
-    functionCall: {
-        name: string;
-        arguments: Record<string, any>;
-    };
-};
-
-/**
- * A function-related message in OpenAI's chat format.
- *
- * @see {@link https://platform.openai.com/docs/api-reference/chat/create#chat/create-messages Create Chat Completion Request Body - messages}
- */
-export interface CreateChatCompletionFunctionMessage {
-    role: Extract<ChatCompletionRequestMessageRoleEnum, "function">;
-    content: string;
-    name: string;
-
-    function_call?: undefined;
-}
-
-export type FunctionMessage = Message & {
-    role: "function";
-    name: string;
-    content: string;
-};
-
-/**
- * A function that can be called by the model.
- */
-export interface CreateChatCompletionFunction {
-    name: string;
-    description?: string;
-    parameters?: JsonSchemaObject;
-}
 
 /**
  * Request body for a Create Chat Completion request, which creates a completion for the conversation's messages.
@@ -386,84 +276,6 @@ export interface CreateChatCompletionStreamResponse {
 }
 
 /**
- * Request body for a Create Moderation request, which classifies if text violates OpenAI's Content Policy.
- *
- * Endpoint: `POST https://api.openai.com/v1/moderations`
- *
- * @see {@link https://platform.openai.com/docs/api-reference/moderations/create Create Moderation Request Body}
- */
-export interface CreateModerationRequest {
-    /**
-     * Your OpenAI API key.
-     */
-    apiKey: string;
-
-    /**
-     * The text to classify.
-     */
-    input: string;
-}
-
-/**
- * Response body for a {@link CreateModerationRequest Create Moderation Request}.
- */
-export interface CreateModerationResponse {
-    /**
-     * The ID of the request.
-     */
-    id: string;
-
-    /**
-     * The moderation model used to classify the text.
-     */
-    model: string;
-
-    /**
-     * The classification results.
-     */
-    results: [
-        {
-            /**
-             * Contains a dictionary of per-category binary usage policies violation flags.
-             * For each category, the value is `true` if the model flags the corresponding category as violated, `false` otherwise.
-             *
-             * @see {@link https://platform.openai.com/docs/guides/moderation/overview Moderation Categories}
-             */
-            categories: {
-                hate: boolean;
-                "hate/threatening": boolean;
-                "self-harm": boolean;
-                sexual: boolean;
-                "sexual/minors": boolean;
-                violence: boolean;
-                "violence/graphic": boolean;
-            };
-
-            /**
-             * Contains a dictionary of per-category raw scores output by the model, denoting the model's confidence that the input violates the OpenAI's policy for the category.
-             * The value is between 0 and 1, where higher values denote higher confidence. The scores should not be interpreted as probabilities.
-             *
-             * @see {@link https://platform.openai.com/docs/guides/moderation/overview Moderation Categories}
-             */
-            category_scores: {
-                hate: number;
-                "hate/threatening": number;
-                "self-harm": number;
-                sexual: number;
-                "sexual/minors": number;
-                violence: number;
-                "violence/graphic": number;
-            };
-
-            /**
-             * Set to `true` if the model classifies the content as violating OpenAI's usage policies, `false` otherwise.
-             */
-            flagged: boolean;
-        }
-    ];
-}
-
-/**
  * Configuration options for streamed Chat Completions executed in dry mode.
  */
 export interface CreateDryChatCompletionConfig {
@@ -488,94 +300,3 @@ export interface CreateDryChatCompletionConfig {
      */
     chunkDelay?: number;
 }
-
-/**
- * **OpenAI-specific** configuration options for the {@link Conversation} class, used by the {@link ConversationConfig} class.
- *
- * @remarks
- * The `messages` property is omitted since messages is handled entirely by the library.
- */
-export type ChatCompletionConfigOptions = Omit<
-    Partial<CreateChatCompletionRequest>,
-    "messages"
->;
-
-/**
- * **Library specific** configuration options for the {@link Conversation} class, used by the {@link ConversationConfig} class.
- */
-export interface ConversationConfigOptions {
-    /**
-     * The first system message to set the context for the GPT model.
-     *
-     * @default "You are a large language model trained by OpenAI. Answer as concisely as possible."
-     */
-    context?: string;
-
-    /**
-     * Dry run. Don't send any requests to OpenAI. Responses will mirror the last message in the conversation.
-     *
-     * @default false
-     */
-    dry?: boolean;
-
-    /**
-     * By default, messages are checked for violations of the OpenAI Community Guidelines and throw an error if any are found.\
-     * Set this to `true` to disable this check.\
-     * Set this to `"soft"` to still check for violations, but not throw an error if any are found. The violations will be added to the `flags` property of the message.
-     *
-     * **Note:** This is not recommended, as it could result in account suspension. Additionally, {@link https://platform.openai.com/docs/guides/moderation OpenAI's Moderation API} is free to use.
-     *
-     * @default false
-     */
-    disableModeration?: boolean | "soft";
-}
-
-/**
- * Listener for when a {@link Message message}'s content is updated.
- */
-export type MessageUpdateListener = (
-    /**
-     * The new content of the message.
-     */
-    content: string | null,
-
-    /**
-     * The {@link Message message} instance that was updated.
-     */
-    message: Message
-) => void;
-
-/**
- * Listener for when a {@link Message message}'s streaming state is updated.
- */
-export type MessageStreamingListener = (
-    /**
-     * The new streaming state of the message.
-     */
-    isStreaming: boolean,
-
-    /**
-     * The {@link Message message} instance that was updated.
-     */
-    message: Message
-) => void;
-
-/**
- * Listener for when a {@link Message message} starts streaming.
- */
-export type MessageStreamingStartListener = (
-    /**
-     * The {@link Message message} instance that started streaming.
-     */
-    message: Message
-) => void;
-
-/**
- * Listener for when a {@link Message message} stops streaming.
- */
-export type MessageStreamingStopListener = (
-    /**
-     * The {@link Message message} instance that stopped streaming.
-     */
-    message: Message
-) => void;
